new function homogenize_gamma
Here I split off the part of gamma_product_simp that globally examines an expression for terms of the form gamma(X + integer) & gamma(X) to a new function homogenize_gamma. This allows the remaining code in gamma_product_simp to simply map over products. We do that by doing a substitution on "*".

I change the name of the top level function from gamma_product_simp to gamma_simp. That's because the homogenization simplification works on more than a product of  gamma expressions.

The homogenize_gamma function includes a new call to facsum. But sometimes further processing expands the result of the facsum.

Update the testsuite & and tidy some other source code comments.