change: map gamma_product_simp over products
Previously we only mapped gamma_product_simp over products.  Maybe this was  maybe efficient, but
doing this fails to simplify, for example, gamma(2*x+1)/2-x*gamma(2*x)
to zero.  So now we apply gamma_product_simp  to the entire expression.

The testing code is in flux, but for now the failures are (11 24 36 40 41 46 51).