additional call to homogenize_gamma
o Set radsubstflag to false. I'm not sure where it happens, but if radsubstflag is true, Maxima sometimes introduces complex exponentials and  unexpected logarithms. It's not wrong, just surprising.

o When the expression changes, apply homogenize_gamma following gamma_product simp. It can happen that this gives an additional simplification.

o From rtest_gamma_simp, remove the assignment to radsubstflag--this option variable is set at the top level of gamma_simp.

o Put in a radcan and allow one test to pass.