move homogenize_gamma up
Move homogenize_gamma to before the
gamma_arg_increase(gamma_arg_decrease(e)).

The function homogenize_gamma decreases the number of gamma functions in an expression, but gamma_arg_increase and gamma_arg_decrease maintains the same number. So it's better to dispatch  homogenize_gamma  earlier.

This changes on test result--the new result has one fewer gammas in it.