gamma_arg_increase and gamma_arg_increase to the last simplification
Move the gamma_arg_increase and gamma_arg_increase to the last simplification. Why? These transformations don't reduce the number of gamma terms, but the other simplifications do.

This change makes several test give better answers. But the answer for
gamma_simp(makegamma(pochhammer(a,n) = (a+n-1)* pochhammer(a-1,n)/(a-1)));

is somewhat worse.